requirements.txt
discord.py>=2.6.0
python-dotenv>=1.0.0
yt-dlp>=2024.0.0
mutagen>=1.47.0